+++ playground/src/app/pages/Table/index.tsx
@@ -1,0 +1,43 @@
+import { BooleanProp, Demo, NumberProp } from "@cianciarusocataldo/demo-ui";
+import { Page, Table } from "modular-ui-preview";
+import {
+  useCommonTranslation,
+  useTableTranslation,
+} from "app/hooks/localization";
+
+export const TableWrapper = () => {
+  return (
+    <Demo
+      label="Table"
+      props={{ Rows: NumberProp(1), Headers: BooleanProp(true) }}
+    >
+      {(props: any) => {
+        let rows = [];
+        if (props.Headers) {
+          rows.push(["Table header 1", "Table header 2", "Table header 3"]);
+        }
+        for (let i = 0; i < props.Rows; i++) {
+          rows.push(["Table cell", "Table cell", "Table cell"]);
+        }
+        return <Table rows={rows} headers={props.Headers} />;
+      }}
+    </Demo>
+  );
+};
+
+const TablePage = () => {
+  const t = useTableTranslation();
+  const tCommon = useCommonTranslation();
+  return (
+    <Page>
+      <p className="text-4xl mt-12 mb-5 ml-3 text-white">{`${tCommon(
+        "molecules"
+      )} - ${t("title")}`}</p>
+      <div className="flex flex-col p-3 lg:flex-row xl:flex-row 2xl:flex-row 3xl:flex-row 4xl:flex-row ">
+        <TableWrapper />
+      </div>
+    </Page>
+  );
+};
+
+export default TablePage;