+++ src/components/molecules/Table/types.d.ts
@@ -1,0 +1,4 @@
+export interface TableProps {
+  headers?: boolean;
+  rows?: (JSX.Element | string)[][];
+}